Update to 0.1.9
Switch from bson to json
Fix I/O bug
Fix filenames
