+++ pyproject.toml
@@ -2,7 +2,6 @@
 requires = [
     "setuptools>=42",
     "wheel",
-    "pymongo>=3.11.0",
     "tifffile==2021.4.8",
     "pillow>=8.2.0",
     "numpy==1.20.3"
